// src/modules/store-catalog/facade/store-catalog.facade.interface.ts
export interface StoreCatalogFacadeInterface {
  find(input: FindStoreCatalogFacadeInputDTO): Promise<FindStoreCatalogFacadeOutputDTO>
  findAll(): Promise<FindAllStoreCatalogFacadeOutputDTO>
}

export interface FindStoreCatalogFacadeInputDTO {
  id: string
}

export interface FindStoreCatalogFacadeOutputDTO {
  id: string
  name: string
  description: string
  salesPrice: number
}

export interface FindAllStoreCatalogFacadeOutputDTO {
  products: {
    id: string
    name: string
    description: string
    salesPrice: number
  } []
}